# [#195] Fix parser exponential behavior