Fixed the exponential behavior, I've had to change a bunch of functions, but in the end it let me refactor `tomlP` nicely so I think it's worth it. Benchmarking also improved.

Since some signatures changed, I've had to modify some tests.

I've also added an `eof` to the TOML parser to make sure the whole `.toml` file is valid.

Benchmarks:
Without fix:
```
benchmarked Parse/tomland
time                 440.8 μs   (430.0 μs .. 453.9 μs)
                     0.993 R²   (0.989 R² .. 0.997 R²)
mean                 433.0 μs   (427.9 μs .. 440.4 μs)
std dev              20.27 μs   (14.42 μs .. 31.25 μs)
variance introduced by outliers: 27% (moderately inflated)
```

With fix:
```
benchmarked Parse/tomland
time                 355.8 μs   (348.9 μs .. 368.5 μs)
                     0.990 R²   (0.980 R² .. 0.998 R²)
mean                 360.6 μs   (356.1 μs .. 368.5 μs)
std dev              19.44 μs   (13.67 μs .. 27.89 μs)
variance introduced by outliers: 32% (moderately inflated)
```
That's a 15% speedup.