@jiegillet 2019-06-23
Some checks are failing: `htoml-megaparsec` doesn't build on `8.4.4`, not sure what I can do about that.