@jiegillet 2019-06-26
Sorry I ended up moving a lot of things around in tests. I ditched all the partial tables and arrays tests in favor of full on `tomlP` tests. I didn't remove any test, and I added one with many arrays.